Validate every game in a multi game YAML
roll_options was rolling the game and then validating that game's
options only. Which means you could get game A validated on upload and
then get game B rolled on generation time which could've contained
invalid options.

This fixes it by validating the YAML for every game, changing all
weights to 0 except for the game it's validating, forcing roll_settings
to roll that one game.