Add Dummy empty world to generate with when checking yamls
The reason for that is that some world (right now I know of jak &
daxter, check options differently when they're in presence of other
games. I could obviously duplicate the YAML and have a multiworld like
that but then generate_early would be twice as slow. By having an empty
world it should basically not impact generation time and still make
worlds thing they're not alone